packaging: fix symlink for snapd.session-agent.socket
The dh_link man-page says that relative syntax is not supported
when describing the source/target in snapd.links. This PR
converts it to absolute names and dh_link will convert it back
to a (short) relative symlink.

It also updates the debian packaging to include the link. It was
forgoten there it seems.

This should fix a failure the broken symlink we currently see
in the snapd snap that prevents the snapd snap from getting
updated.