tests: add test for snapd.session-agent.socket symlinks
Rename "debs-have-built-using" to a more generic "debs" test
and also test there that the symlink for snapd.session-agent.socket
in the package points to the right place. There was a bug in
the packaging that resulted in a bad symlink of the form:
```
$ dpkg -c snapd_2.40+git1432.803f873~ubuntu18.04.1_amd64.deb |grep snapd.session-agent.socket
-rw-r--r-- root/root       152 2019-08-15 16:16 ./usr/lib/systemd/user/snapd.session-agent.socket
lrwxrwxrwx root/root         0 2019-08-15 16:16 ./usr/lib/systemd/user/sockets.target.wants/snapd.session-agent.socket -> /snapd.session-agent.socket
```
This commit ensures this won't regress.