Add /var/cache/snapd to the snapd state
Now the catalog is not gonna be refreshed on each test due to the
directory /var/cache/snapd is already included in the snapd state which
is saved during the project preparation.

This change is done to avoid the error 429 sent from the store which
indicates there were too many requests.