tests: allow test user XDG_RUNTIME_DIR to phase out
On Arch it seems that using sudo / su for working with the test user
spawns the /run/user/12345 tmpfs for XDG_RUNTIME_DIR which asynchronously
cleans up itself sometime after the test but not instantly, leading to
random failures in the mount leak detector. Give it a moment but don't
clean it up ourselves, this should report actual test errors, if any.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>