httputil: rework protocol error detection
Our code tries to detect http2 PROTOCOL_ERRORS and retries the
request in this case.

Unfortunately it seems this is not easy to detect. In 42dc19a this
code tried to be smart and detect this via http2.StreamError but it
seems like with the h2_bundle.go in the go distro this does not
work, i.e. in https://travis-ci.org/snapcore/snapd/jobs/575471665
we still got protocol errors even with this detection code.

So this code falls back to simple and naive detection.