tests: move restore-project-each code to existing function
We have a restore_project_each function that contains various existing
logic for restoring. We call it from restore-each spread stanza.
Just after that call there a bit more code that checks the journal
for signal-based termination messages and SELinux mislabelling.
This code belongs in the function so move it there for consistency.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>