patch: add an overlord patch that normalizes channel in state
This PR adds a patch that will normalize the channel in the state.
Snapd was not normalizing the channel so it is possible that the
state contains channels like `/stable` or even `//edge//` without
perceivable ill effects. However with the upcoming change to
"default-tracks" that can be set store side the semantic of those
become less clear. To avoid any ambiguity we normalize the channels
so "/edge" -> "edge".