daemon/api_model_test.go: always unlock snapState
If for example st.Get() returned non-nil error, we would exit the function while
still holding the lock and cause further tests to block indefinitely trying to
lock an already locked mutex.

Note that we can't just defer unlocking the state because getSerial below in the
test needs to lock the mutex, so instead we need to manually unlock it before
continuing, this defered if + unlock ensures we always unlock appropriately.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>