tests: add --quiet switch to retry-tool
Pawel made a case for using retry-tool in situations where we know and
anticipate the first few attempts of a command to fail, e.g. while
waiting for some asynchronous operation to finish. In that case the
default output doesn't help much, as we routinely fail on the road to
eventual success.

For this use case, add the --quiet switch, to silence all output from
the tool itself.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>