tests: fix mountinfo-tool filtering when used with rewriting
The mountinfo-tool program has an ability to rewrite certain
non-deterministic aspects of the mount table, such as revisions in paths
like /snap/core/1234

This inadvisedly clashed with the filtering logic that allows to look
at a specific path. The rewrite would cause the filter not to match.

This patch adds a new property to each mountinfo entry object and changes
the filtering and display code so that filtering is applied early, before
rewrite. Instead of discarding elements we filtered out we simply set
the matching attribute and not display them. This is done so that
we can continue to rewrite exactly as before, producing identical
rewritten results.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>