cmd/snap-confine: fix /snap duplication in legacy mode
While working on MS_SHARED tests I realized the new test was failing on
ubuntu-core-16 when checking a particular property of a core-16 base
snap.  In snap-confine terms, this combination triggers legacy mode,
where the snap mount namespace is crafted using / as the root
filesystem. This is in contrast to normal mode where the base snap is
the root filesystem.

The particular test in question was failing because, it seems, certain
thins were mounted twice. In particular, all the snaps were mounted
twice.

This can be checked by looking at the data file for the
tests/main/mount-ns test, which for google:ubuntu-core-16 backend, for
the PER-SNAP-16 data file, shows the following content:

    1073 1071 0:2 / /snap/core/1 ro,nodev,relatime master:60 - squashfs /dev/loop2 ro
    1074 1072 0:2 / /snap/core/1 ro,nodev,relatime master:60 - squashfs /dev/loop2 ro
    1075 1071 0:3 / /snap/core18/1 ro,nodev,relatime master:61 - squashfs /dev/loop3 ro
    1076 1072 0:3 / /snap/core18/1 ro,nodev,relatime master:61 - squashfs /dev/loop3 ro
    1077 1071 0:4 / /snap/pc-kernel/1 ro,nodev,relatime master:62 - squashfs /dev/loop4 ro
    1078 1072 0:4 / /snap/pc-kernel/1 ro,nodev,relatime master:62 - squashfs /dev/loop4 ro
    1079 1071 0:5 / /snap/pc/1 ro,nodev,relatime master:63 - squashfs /dev/loop5 ro
    1080 1072 0:5 / /snap/pc/1 ro,nodev,relatime master:63 - squashfs /dev/loop5 ro
    1081 1071 0:6 / /snap/test-snapd-mountinfo-core16/1 ro,nodev,relatime master:64 - squashfs /dev/loop6 ro
    1082 1072 0:6 / /snap/test-snapd-mountinfo-core16/1 ro,nodev,relatime master:64 - squashfs /dev/loop6 ro
    1083 1071 0:7 / /snap/test-snapd-mountinfo-core18/1 ro,nodev,relatime master:65 - squashfs /dev/loop7 ro
    1084 1072 0:7 / /snap/test-snapd-mountinfo-core18/1 ro,nodev,relatime master:65 - squashfs /dev/loop7 ro
    1085 1071 0:8 / /snap/test-snapd-rsync/1 ro,nodev,relatime master:66 - squashfs /dev/loop8 ro
    1086 1072 0:8 / /snap/test-snapd-rsync/1 ro,nodev,relatime master:66 - squashfs /dev/loop8 ro

The file in question shows the per-snap mount namespace of an
application running in legacy mode. This is in stark contrast to the
PER-SNAP-18 data file which shows just:

    1071 1070 0:2 / /snap/core/1 ro,nodev,relatime master:60 - squashfs /dev/loop2 ro
    1072 1070 0:3 / /snap/core18/1 ro,nodev,relatime master:61 - squashfs /dev/loop3 ro
    1073 1070 0:4 / /snap/pc-kernel/1 ro,nodev,relatime master:62 - squashfs /dev/loop4 ro
    1074 1070 0:5 / /snap/pc/1 ro,nodev,relatime master:63 - squashfs /dev/loop5 ro
    1075 1070 0:6 / /snap/test-snapd-mountinfo-core16/1 ro,nodev,relatime master:64 - squashfs /dev/loop6 ro
    1076 1070 0:7 / /snap/test-snapd-mountinfo-core18/1 ro,nodev,relatime master:65 - squashfs /dev/loop7 ro
    1077 1070 0:8 / /snap/test-snapd-rsync/1 ro,nodev,relatime master:66 - squashfs /dev/loop8 ro

The cause of this duplication is snap-confine's $SNAP_MOUNT_DIR mapping
code which brings the real snap mount directory, perhaps
/var/lib/snapd/snap from the host mount namespace to /snap in the
per-snap mount namespace.

When invoked in legacy mode, /snap is already populated, as we are
running on top of a copy of the real initial mount namespace, so we
essentially were doing the equivalent of:

    mount --rbind /snap /snap

This patch makes the mapping code condition and only active in normal
mode. The data sets for mount namespace tests were updated to reflect
this.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>