tests: extend mount-ns test to handle mimics
The mount namespace test is very useful for exploring the layout of the
mount namespace used by snapd under various circumstances. The test
already covered several key combinations of the mount namespace: Ubuntu
16.04 and 18.04 both as classic and core hosts also as base snaps. The
test explored the vanilla behavior of the mount namespace, as
constructed by snap-confine.

This patch extends that test to cover the properties of writable mimic,
as created by snap-update-ns. This test is meant as a prerequisite for
the upcoming change to propagation in mount namespaces.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>