timeutil: fix schedules with ambiguous nth weekday spans
The nth weekday is confusing on it's own. Adding week spans, the end result is
fuzzy at best.

Consider the following calendar:
        August 2018
   Su Mo Tu We Th Fr Sa
             1  2  3  4
    5  6  7  8  9 10 11
   12 13 14 15 16 17 18
   19 20 21 22 23 24 25
   26 27 28 29 30 31

The schedule `mon1-tue1` matches 06.08 and 07.08, both days being he first
occurrence within the month.

However, extending the schedule to `mon1-fri1`, this should match: 01.08-03.08
and 06.08-07.08 (5 days total), so that the weekdays are inside the week span
defined in the schedule. If we let go of this property, and just match anything
between 03.08 and 06.08, the update could happen over the weekend, which is
clearly not intended.

Similarly, a schedule `mon2-fri2`, matches dates: 13.08-14.08 and 08.08-10.08 (5
days total).

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>