tests: remove mount_id and parent_id from mount-ns test data
The mount-ns test is useful as it shows a realistic view of the mount
table. The mount IDs are re-numbered and are pretty much just a
sequence. While some work went into making it easy to understand which
mount namespace we are looking at, by checking the mount identifier
value for membership in ranges 0..999, 1000..1999 or 2000..2999 - the
overall value comes at a great cost. Any insertions or removals cause
massive re-numbering to occur below the change. This makes it difficult
to evaluate differences that a patch is introducing without manually
convincing oneself of the "real" diff buried inside the diff printed by
the tooling.

This patch discards the first fro columns, namely mount_id and parent_id
in the hope that the resulting output is clearer and more useful.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>