tests/mountinfo-tool: proper formatting of opt_fields
The code for formatting optional fields, those that are used to convey
information about mount namespace sharing options, was a special-case.
When mountinfo-tool was invoked without attribute specific formatting,
it would print optional fields correctly. If .opt_fields was passed on
command line the formatting would differ, because the fields were really
a list. There was some special handling that would make it look nicer
but it was still different from the case without attribute selection
because empty list was not formatted as "-". This made it difficult to
use .opt_fields reliably for formatting output.

This patch fixes this by adding a new type, OptionalFields, that has
appropriate __str__ method. All the special cases were removed,
simplifying remaining code.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>