cmd/model: only output brand for snap model w/o opts, in verbose do brand-id
With `snap model --serial`, `snap model --verbose` and
`snap model --verbose --serial`, we want to output the normal brand-id because
that's an immutable and expected value, while the brand output for model is
stylized and may be displayed slightly differently depending on the situation.

Also edit the unit test for this change.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>