usersession/userd: make sure to export DBus interfaces before requesting a name
The problem was reported by a user in the forum[1]. When attempting to open a
URL from the snap application, it fails on the first attempt. This happens when
userd is launched as a result of DBus request.

In the `snap run --shell`, the problem is logged like this:

```
[maciek@sloop maciek]$ xdg-open http://google.com
user-open error: Object does not implement the interface
[maciek@sloop maciek]$ xdg-open http://google.com
[maciek@sloop maciek]$
```

The actual reason is a a race between DBus, which has queued the call message,
and the interface becoming exported (and thus handled) at godbus level. When
there is no handler for given call at godbus level, it returns a generic error
message which is related back to the called. The problem is also reproducible
with a dbus-send call.

1. https://forum.snapcraft.io/t/xdg-open-does-nothing-on-first-invocation/13278

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>