gadget: do not fail the update when old gadget snap is missing bare content
It was brought to our attention, that there exist gadget snaps in the wild,
where gadget YAML lists content that does not exist within the gadget snap. In
such scenario, the update would fail on the attempt to layout the old volume.

However, for the purpose of update process, we only need to identify the
structures that are to be updated. Because we backup whatever is changed during
update, we do not care much about the content of the old gadget snap. Apply more
relaxed layout processing and do not verify that the data existed int he old
gadget snap.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>