timeutil: fix schedule ambiguity, add anchor points, make schedule a contigous 1 week span
As agreed, in order to resolve the ambiguity, each schedule is anchored at
either start (eg. mon1-fri), or at the end (mon-fri1). The anchor defines the
fixed time-point of the schedule, where the other end defines the span around
that point. Previously accepted schedules are not silently degraded to be
anchored at the starting point.

Having resolved the ambiguity, a week span including the last day of the month
can now be accepted and handled correctly, thus mon5-fri or mon-fri5 are valid schedules.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>