cmd/snap-confine: simplify hiding /writable
Initial implementation of this idea used a loop that detached anything
under /writable or /var/lib/snapd/hostfs/writable. The loop needed
because it was hard to determine what needed to be unmounted and some of
the changes propagated so we were always attempting to unmont the first
entry we saw and then tried again, to see if there's anything left to
do.

This patch significantly simplifies this logic. There are two ideas
employed:

1) Instead of using a loop, we can ask the kernel to detach
/var/lib/snapd/hostfs/writable *recursively*.

2) Depending on the version of core used as well as the mode in which we
are executing we may need to detach /writable in addition to
/var/lib/snapd/hostfs/writable. Specifically /writable is only present
in legacy mode on core16.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>