cmd/snap-confine: allow digits in hook names
Interface hooks react to interface connection events. Interface names
can contain digits, for example "i2c" is a valid plug or slot name.

It was reported that interface hooks for the i2c interface were not
working. This was traced to an overly-strict validation of the snap
security tag computed for the hook by snapd, performed by snap-confine.

Fixes: https://bugs.launchpad.net/snappy/+bug/1845636
Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>