osutil: drop error return path from StreamEqual
I'm about to merge StreamEqual and StreamsEqual. They currently differ
in signature: one takes an buffer size and returns an error in addition
to boolean, the other does not.

I decided to drop the error value as we don't really need it in the
current scheme. Just the equal-or-not value.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>