client: refactor the timeout to be per single raw request
The higher level client.do() call already sets up a timeout that covers the
retries of client.raw() requests. Add a helper to up a per-request timeout that
covers both the request and the response. Since the timeout is set for a single
request only, we can properly distinguish the errors caused by reaching the
request deadline and ones caused by other factors (eg. no socket availble) that
trigger the retry mechanism.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>