tests/main/searching: limit the number of calls to snap find
During tests, snapd is run with SNAPD_DEBUG_HTTP=7 which causes a lot of logging
to stderr. The logs are forwarded to journald and later end up on disk.

On openSUSE 15.1, we have observed that excessive logging, especially for snap
find --section=featured, may choke journald and cause the whole of the API call
to snapd take much longer than expected.

Specifically a request like this was observed:
Oct 02 06:36:47 oct020627-523151 snapd[8948]: retry.go:61:
    DEBUG: The retry loop for https://api.snapcraft.io/api/v1/snaps/search?...&section=featured
    finished after 1 retries, elapsed time=1m0.45638716s, status: 200

Try to workaround the problem by reusing cached test data.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>