cmd/model: don't show model with display-name inline w/ opts
With snap model we want to show the display-name with the actual model in paranthesis for easy operator (i.e. human) inspection, but with snap model --verbose we want the model key to be the verbatim model so it is easy to inspect the output programatically in scripts and such.

Also add a unit test for --verbose output when there's a display-name in the model assertion.

Also fix a typo in the diff output for one of the spread tests in the error condition, and simplify the spread test to always use --verbose to ensure that the model in the model assertion from snap known matches that from snap model.