daemon: change /v2/download API to take "snap-name" as input
This changes the /v2/download API as discussed during the standup
and on irc to take a single "snap-name" and no action.

No backward compatiblity code is provided. This should be ok
because the interface was never advertised and we don't have any
matching client support yet.

Providing backward compatiblity is trivial and I am happy to
add it (requires a bit more testing code though).