test-snapd-netplan-apply: add info command to the snap
The new version of Netplan also exposes an "Info" method, so add an app to the
test netplan snap to access this (since the netplan binary won't by default use
the Info method at all.

Also tie the version of the snap to the version of the netplan part in the
snapcraft using override-pull and some snapcrafters git magic.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>