.gitignore: scope "share" to project root
The "share" entry is problematic as it can surprisingly block
usr/share/foo that some test snaps contain. The origin of the share
entry indicates it is related to git-buildpackage. I don't know any
actual users of this anymore but for the sake of not breaking them, just
make the entry more precise.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>