spread: generate delta when using google backend
We never generated a delta when using google backend. This has an adverse effect
on the size of the source package being uploaded, and ends up being super
annoying on links with slow upload speed.

The patch enables delta generation for google (incl. google-unstable) and linode
backends.

In my testing, the size of the uploaded package went down from 5.5MB to 1.2MB.

While at it, the reference snapd revision is bumped to 2.42 tag.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>