snapstate: add LastActiveDisabledServices to SnapState
This field will be used for saving disabled services in the following
situations:

* When we have _removed_ systemd units for a snap during a refresh, revert, or
  disable and thus systemd can no longer remember if the service was disabled or
  not.
* When we have refreshed, or reverted to a revision of the snap where a service
  that was previously present and disabled is no longer present in the snap and
  thus can't be disabled.

Note we need the second situation because if one disables a service, then
refreshes to a revision where the service was renamed, and we revert back to the
previous revision due to some problem, that previously disabled service will be
erroneously enabled.

This is specifically _not_ a current list of what services are enabled/disabled.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>