gadget: helper for volume compatibility checks
Add a helper for checking volume and structure compatibility, with primary use
in a remodel scenario.

The compatiblity check performs only a partial layout, thus missing content
cannot be detected at this stage.

While at it, make sure that structure name transition is only allowed when using
a MBR partitioning schema. In case of a GPT one, the structure name is written
to the GPT as PARTLABEL.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>