gadget: rename existing and add new helpers for checking filesystem/parition presence
The IsBare() helper that checks whether a structure has a filesystem or not is
confusing, since structures without a partition table entry are also referred to
as 'bare'. Rename to helper to make the substance of its check more explicit. At
the same time, add a new helper that checks whether a structure describes a
partition.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>