interfaces: allow introspecting network-manager on core
We were alerted about a program that fails to work on core systems, with
the network-manager plug connected, because said program used
introspection to discover the available DBus methods.

Introspection was not a part of the network manager interface before.
After discussion with security is is now conditionally added on core
systems only, where the peer is scoped to a snap.pkg.app label that
pinpoints a specific snap. On classic system that would be unavoidably
"unconfined" where it would apply to all the classic services.

Fixes: https://bugs.launchpad.net/snapd/+bug/1849291
Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>