spread: enable bboozzoo/snapd-devel-deps COPR repo for getting golang-x-xerrors
The golang-x-xerrors package is available for Fedora 31 and later. While we have
filed a ticket https://bugzilla.redhat.com/show_bug.cgi?id=1764123 and reached
out the maintainers to get a build for Fedora 30 pushed to the repositories, it
may take a while for it to be resolved.

In the meantime, I've setup a COPR repo bboozzoo/snapd-devel-deps that contains
the required dependency and could be used in future when similar issues appear.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>