cmd/snap: support (but warn) using deprecated multi-slash channel
Before, when `--channel=<foo>` was just passed on to the store, we
inherited a bit of undefined behaviour where
`--channel=//foo/////bar//baz/` was interpreted the same as
`--channel=foo/bar/baz`. In particular, somebody doing
`--channel=$FOO/candidate` in a script could rely on it working even
if `$FOO` was `""`.

The change to support `--channel=<risk>` meaning `--channel=<current
track>/<risk>` means that we now parse `--channel`, and lost this bit
of behaviour.

Now, we don't _want_ this behaviour: it presents a problem for the
introduction of default tracks (coming RSN!). But we dont' want to
break our users. So...

This change makes it so that the old spuriously-slashed channels are
accepted, and warned about. The warning will go away in a couple of
releases when we start enforcing clean channel names.

This change does _not_ address having spuriously-slashed channels in
state (which currently leaves you in a very strange situation). That
will come in a separate PR.