interfaces/content: workaround for renamed target
This patch works around a bug in misuse of the content interface in
direct mode when multiple slots are connected to one plug. For context
please follow the text below.

The content interface supports two fundamental modes: direct and spool.

The first mode, direct, is the oldest, where a directory is bind-mounted
from one place to another. This mode only works when the cardinality of
the set of mount sources over the mount target is equal to one. To use
this mode one has to define a pair of snaps with the following interface
declarations:

    interfaces:
	content:  # slot
	    read: $SNAP_DATA/source

    interfaces:
	content:  # plug
	    target: $SNAP_DATA/target

This fragment will correctly bind-mount source at target.

The second mode, spool, allows using multiple sources in one target. In
this mode the target (plug) becomes a spool directory and each source,
either read or write becomes a element in the spool directory. To use
this mode one has to define a slot with the "source" attribute as the
parent for either "read" or "write" attributes.

    interfaces:
	content:  # slot
	    source:
		read: [$SNAP_DATA/source]

This fragment will correctly bind mount source at target/source. Careful
reader might notice that this presents a new problem. What might happen
if multiple slots all want to provide a directory called "source". Such
declarations would over-write each other, defeating the purpose of
the new mode for app-with-plugins style of usage.

When the spool mode was introduced this problem was identified and the
mount backend was extended to automatically resolve such conflicts by
altering the name of the mount target from "target/source" to
"target/source-2" (for the second element that wishes to provide
"source" to "target").

The error reported as https://bugs.launchpad.net/snapd/+bug/1849845
arises when two or more slots are connected to one plug in direct mode.
The mount backend creates additional "target-2", "target-3" directories
but the apparmor backend is not matching that with appropriate
permissions.

Historically each security backend was entirely standalone and separate.
In practice, ever since the snap-update-ns apparmor profiles were
introduced, the apparmor and mount backends must agree on the
directories that need to be created and mounted. This is only done
implicitly for now.

The patch adjusts the apparmor specification for snap-update-ns so that
it can create and mount directories named "target{,-[0-9]*}/", in
anticipation of possibility of renaming.

This problem intersects with validation, connection cardinality and
upcoming work on greedy plugs and slots. Some improvements to usability
are possible once all are taken into account.

Fixes: https://bugs.launchpad.net/snapd/+bug/1849845
Fixes: https://bugs.launchpad.net/stuttgart/+bug/1849804

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>