tests/overlord/snapstate: delete LastActiveDisabledServices from some tests
It's no longer needed here since we specify omitempty for the json field
annotation and as such shows up as nil instead of an empty (but non-nil) list.
Adding it here in the first place was the wrong thing to do as well, as
additional fields ideally should be ignored in other tests like this where they
don't get set.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>