cmd/snap: wait while inhibition file is present
Inhibition files are stored in /run/snapd/inhibit and are named after
the snap name. Presence of a non-empty file stalls "snap run" from
starting additional application processes or hooks.

A non-empty file contains a hint about what is going on that doesn't
require a round trip to snapd.

Zenity, text-based and headless UIs are provided while inhibition is in
place. Currently it shows a simple progress bar and doesn't talk to
snapd. Eventually it can ask snapd for inhibition details via a REST
API.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>