features: add ClassicPreservesXdgRuntimeDir feature
The new feature controls the value of XDG_RUNITME_DIR for snaps using
classic confinement. When the feature is enabled classically confined
snaps see the regular value of XDG_RUNTIME_DIR, like all other non-snap
software. When the feature is disabled (which is the case by default)
then the XDG_RUNTIME_DIR is changes the same way as it is done for snaps
using strict confinement.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>