sandbox/seccomp: accept build ID generated by Go toolchain
From user reports, it is possible that snap-seccomp binaries found in the wild,
will contain build ID generated by Go toolchain. The build ID in this case is a
hash of go package, import path, binary, then truncated to 83 bytes and written
to .note.go.buildid ELF section.

See https://aur.archlinux.org/packages/snapd/#comment-714427 and
https://forum.snapcraft.io/t/future-release-to-include-alpine-linux-as-snapd-host/13144/7
for examples.

Update the code to allow the upper limit of 166 characters, which is the
hexstring representation of Go build ID. Since build ID is effectively user
controlled during the build time, this also bumps the upper sanity limit on the
length of this field.

Fixes: https://bugs.launchpad.net/snapd/+bug/1850914

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>