tests/lib/state: snapshot and restore /var/snap during the tests
The /var/snap directory was not properly restored after each test, thus data
that was created during the test would just leak from one test to another. In
this particular case, installation of a snap would create directories under
/var/snap/<snap>/{rev,current}, that would be left behind after restore
completed. Incidentally, this could break snap removal in subsequent tests if
the revisions that were installed before were not accounted for.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>