overlord: mock device serial in gadget remodel unit tests
Device serial number was not mocked properly, what caused the request-serial
task to be injected in the gadget remodel unit tests. This in turn made affected
unit tests run much longer, as the task would timeout and fail. As a secondary
effect, the tasks might not have completed within the settle timeout.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>