cmd/snap-failure: fallback to snap from core, extend tests
The snapd snap has no previous revision if this is the first time it gets
installed. This can happen in a core to base remodel scenario, when the snap
gets installed before all other snapd taking part in a remodel.

Since there is no previous revision of snapd snap that we can call, fall back
and try to run snapd from the core snap.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>