boot,bootloader: make MakeBootable() uc20 aware
The current MakeBootable() code is pretty specific to UC16/UC18.
For UC20 we need to tweak it to install the recovery boot config
in boot.MakeBootable().

This commit adds a new "Recovery" flag to the bootloader.Options.
Based on this flag we will need to be able to look at different
base directories for grub (and potentially the other bootloaders)
as UC20 will normally have /boot/grub (system-boot) but EFI/ubuntu
(system-seed) for the recovery bootloader.

This will also allow to access the recovery bootloader config
via: `bootloader.Find("/run/mnt/ubuntu-seed", Options{Recovery:true})`.
This will be needed in e.g. PR#7762 where after the install from the
emphemeral mode we need to set the bootloader config of the
recovery grub to `snapd_recovery_mode=run`.