cmd/snap-update-ns: ignore EBUSY when removing a directory
Ignoring this error only makes sense when the new robust mount namespace
update feature is enabled. When this feature is enabled a mount
namespace profile is entirely removed before attempting to set up a new
profile.

This allows us to recover from the case of nested mimic, that are
mounted over one another.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>