interfaces/browser-support: allow reading status of huge pages
Recent chromium content api snaps are checking the status of huge pages.
While this (will be) allowed in the hardware-observe interface, also
include it here so browser snaps don't have to plugs it and get more
access than required. Importantly, this access only allows checking the
status of huge pages, not altering the use of them. Allow querying both
the new transparent hugepages and falling back to the older hugetlbfs.

References:
- https://www.kernel.org/doc/Documentation/vm/transhuge.txt