devicestate: ensure system installation
Core20 systems start with only the seed partition on disk and should
be properly installed in order to boot in run mode.  After the system
boots in ephemeral mode and is properly seeded, we should invoke the
install handler. This commit adds infrastructure to invoke this
handler, actual handler contents will be added at a later moment.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>