snap-bootstrap: support auto-detect device in create-partitions
The current code in `snap-bootstrap create-partitions` expects
that we pass a device. This is not really needed because the
device to use can be figured out from the gadget-dir. I.e. it's
enough to find the "system-seed" partition and get the device
from that.

This means the layers above partition creation do not need to
be in the business of finding the right device and all low-level
handling is in the some (code) area.