HACKING.md: add nvidia options to configure example
This is useful for beginners who may not know about the tricky mounting that we
do in snap-confine to enable graphical applications to work correctly with
drivers and libraries from the host's filesystem.

Without these options, if you have nvidia graphics and try to run graphical snap
applications with your local version of snap-confine things will generally not
work well.

Thanks to @degville for the suggested wording.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>