systemd: use an interface for systemctl errors
Tests frequently mock systemctl via MockSystemctl() or plug a command via
MockCommand(). When using the former, it is not possible to properly mock the
state of IsEnabled() call, since a service is disabled when systemctl exits with
status 1 and prints out "disabled" to its stderr.

Converting the low level error returned by systemctl wrapper to an interface,
the tests can now properly mock the desired behavior.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>