tests: remove "test-snapd-tools" in smoke/sandbox on restore
In the autopkgtest environment we do not run any suite prepare
or restore scripts. The smoke tests run on a pristine environment
and the spread.yaml is generated from the d/tests/integrationtest
script. This means that each test must remove any snaps it installed
in the "restore" stanza.

This fixes a rather strange error in the autopkgtest run on 20.04
where we have a test-snapd-sh installed from the "sandbox" suite.
But this test-snapd has a different interface than the test-snapd-sh
from the store.