snap-bootstrap: update encryption test to use uc20 gadget
We were using a previous version of the pc gadget in our spread
tests, so update our encryption test to use the uc20 gadget. This
also exposed us to some scenarios that don't happen in our only
use case, such as creating a seed partition or attempting to
add partitions to a disk with encrypted partitions already added.
Addressing the first case is trivial so we fixed it here, the
other is still not possible and can be fixed in a forthcoming
PR if necessary.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>