store, overlord/snapstate, etc: SnapAction now returns a []…Result
SnapAction will in some situations return a "redirect channel" which
snapstate et al should use to know that something interesting has
happened.

Rather than add this field to *snap.Info, it seemed to me to make more
sense to make SnapAction return it separately.

Rather than make SnapAction return []*snap.Info, []string, error, I
introduce SnapActionResult which contains a *snap.Info for now, and
can easily grow to have a string.

There already exists a snapActionResult used for marhsalling to the
store; hopefully the use is just the right intersection of different
enough and similar enough for it not to cause problems.