boot,makebootable: turn on UC20 kernel extraction and boot state tracking
This turns on boot state tracking for UC20, as well as the kernel asset
extraction for UC20.

Specifically, we will extract the kernel assets to ubuntu-boot partition and
setup the kernel.efi symlink there when preparing for run-mode, and then use
bootState20 for implementing MarkBootSuccessful and SetNextBoot when we have a
modeenv available from the device.

Also add many unit tests around this.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>