data/selinux: workaround incorrect fonts cache labeling on RHEL7
The /usr/lib/fontconfig/cache directory should be labeled as
fonts_cache_t (there's even a file type defined for that in the reference
policy), but isn't on some older systems, eg. F29 and RHEL7. We have added a
workaround for fontconfing cache back in F29 days and poked a hole for
snappy_t (snapd context).

We now need a similar whitelist rule for snappy_mount_t (snap-update-ns) to
allow mounting a pre-made cache from the host.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>