tests: remove revision leaking from ubuntu-core-refresh
This is the second of the three, well-known tests, that leak a snap
revision, which upsets the mount leak detector, which also upsets the
mount-ns test.

This case is simple, since core devices do not remove unused revisions
across tests, we just have to do it manually. There's some additional
waiting involved, so that ongoing changes are completed, as otherwise
removal will fail.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>