gadget: skip update when mounted filesystem content is identical
Even though the structure edition is bumped, the actual content of filesystem
structure may be unchanged. This can only be determined once we run the
`Backup()` phase of the update sequence. Should such scenario be identified,
return an error indicating that there is no update. The callers up the stack are
already prepared to handle such scenarios and trigger no reboot.

The mechanism is already part of the 'raw' structure updater, the patch
introduces it for the filesystem updater.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>