osutil: rewrite same-dir detection in AtomicRename
AtomicRewrite optimizes some of the cost of fsync by detecting if the
source and the target directory of the rename process is the same. This
used to be done by string comparison but this was partially racy since
the directories could have been renamed between opening the first and
second directory.

This patch changes the logic to open both locations and compare the
device and inode numbers.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>