daemon: refactor create-user to a user action & hide behind a flag
This adds POST method support to /v2/users, with an action of "create"
that does the same thing as /v2/create-user (with the latter to be
deprecated later), and an action of "remove" that is to be done in an
upcoming PR.

It also makes all system user admin actions depend on a flag. This
flag is release.OnClassic for now, but could trivially be made to
check a config value if needed.