tests: build the initramfs + kernel snap for UC20 spread tests
This creates a focal debootstrap environment to build the kernel snap
in, which itself requires a debootstrap chroot to install the necessary
packages. We need to do this, because changes to snap-bootstrap need to
be tested, but snap-bootstrap is embedded into the initramfs, which is
in the kernel snap, thus we need to rebuild the kernel snap with our
snap-bootstrap.

Signed-off-by: Ian Johnson <person.uwsome@gmail.com>