many: don't use kernel from the modeenv anymore
This is in preparation for the modeenv to merely track what list of kernels are
valid to have booted from. For now, we shouldn't be using the Kernel setting in
the modeenv at all.

Note that this is because `snap-bootstrap initramfs-mounts` currently determines
what kernel snap to use from the symlinks on the bootloader rather than from the
modeenv - the modeenv will ultimately just be a verification that the kernel 
from the bootloader is trusted.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>