daemon, store: support resuming downloads
This implements a first pass at the snapd side of resumable-via-snapd
downloads.

There's a little dance the client needs to do to find out what it's
going to download: right now snapd replies with the sha3, and the
client should pass that to snapd when wanting to resume. It's not
prefixed but probably should be, so we can switch to using hmac
without having to add another blob field in the api.

It goes like this: first, get the hash and other info,

    $ sudo curl -D- --unix-socket /run/snapd.socket -d '{"snap-name":"http","no-body":true}' http://localhost/v2/download
    HTTP/1.1 200 OK
    Content-Disposition: attachment; filename=http_29.snap
    Content-Type: application/octet-stream
    Snap-Length: 2572288
    Snap-Sha3-384: bdbcbdac7e8701228b01381b4774419465b4bfbed489f0104f44e1ad691b983bfe05b1c60dd46f3c802b0cc3b3c800ea
    Date: Sat, 01 Feb 2020 02:07:30 GMT
    Content-Length: 0

then you ask for the whole thing,

    $ sudo curl -s -D- -o http_29.snap --unix-socket /run/snapd.socket -d '{"snap-name":"http","resume-stamp":"bdbcbdac7e8701228b01381b4774419465b4bfbed489f0104f44e1ad691b983bfe05b1c60dd46f3c802b0cc3b3c800ea"}' http://localhost/v2/download
    HTTP/1.1 200 OK
    Content-Disposition: attachment; filename=http_29.snap
    Content-Length: 2572288
    Content-Type: application/octet-stream
    Snap-Length: 2572288
    Snap-Sha3-384: bdbcbdac7e8701228b01381b4774419465b4bfbed489f0104f44e1ad691b983bfe05b1c60dd46f3c802b0cc3b3c800ea
    Date: Sat, 01 Feb 2020 02:10:41 GMT

then you truncate the file

    $ sudo truncate -s 1MB http_29.snap

and ask to continue (note in 16.04's curl at least the -C option
doesn't work well with POSTs -- it tries to continue the POST? I have
no idea),

    $ sudo curl -s -D- -H Range:bytes=1000000- -o http_29.snap --unix-socket /run/snapd.socket -d '{"snap-name":"http","resume-stamp":"bdbcbdac7e8701228b01381b4774419465b4bfbed489f0104f44e1ad691b983bfe05b1c60dd46f3c802b0cc3b3c800ea"}' http://localhost/v2/download
    HTTP/1.1 206 Partial Content
    Content-Disposition: attachment; filename=http_29.snap
    Content-Length: 1572288
    Content-Range: bytes 1000000-2572287/2572288
    Content-Type: application/octet-stream
    Snap-Length: 2572288
    Snap-Sha3-384: bdbcbdac7e8701228b01381b4774419465b4bfbed489f0104f44e1ad691b983bfe05b1c60dd46f3c802b0cc3b3c800ea
    Date: Sat, 01 Feb 2020 02:16:31 GMT

(also note this change only has snapd supporting ranges like this one,
ie "bytes=<start>-"; everything else results in a 200).

I haven't had time to write unit tests for this, other than for some
of the store bits which are better covered than before (but need
more). Notably the 200 vs 206 thing I describe below is NOT covered.

I've tested it "live" using curl and httpie, and it's sensible and
works, and can resume downloading a snap that's in the cache.  Right
now the store isn't ever replying with a 206 when I ask for a range,
so I haven't been able to end-to-end that path, but it's not anything
I'm doing wrong AFAICT; the store also doesn't 206 when curl or httpie
talk to it directly. I think this is a bug in the store, but it's also
correct behaviour AFAIK (from reading the RFC).

The client will need to check whether a ranged request comes back with
a 206 or a 200 to know whether it's actually resuming or not.

NOTE the store itself (`downloadImpl`) doesn't support this behaviour;
this needs fixing as currently it'll be downloading the whole thing
twice as soon as it tries to resume once and gets the full content
instead (the sha3 won't match). I tried to dive into fixing this but
it's waaay too late.